parametros se guardan en las funciones correctamente, guardar clases en otro archivo, borrar codigo innecesario
